Handle SASL Mechanisms that do not report negotiated QOP correctly
The way Hadoop handles SASL Mechanisms that do not report the
negotiated QOP after the SCRAM patches is badly broken

Previously, Hadoop implicitly required that the Mechanisms
sets the negotiated QOP, and failed with an NPE otherwise.

With the SCRAM patches, Hadoop will ONLY check the negotiated QOP
if it was set by the mechanism, but omit the check
if the mechanism didn't set anything.

We could discuss whether Mechanisms that accept non-auth QOPS,
but do not report any QOP are buggy or not, but the curent Hadoop
implementation uses the negotiated value in shouldWrap() function
to determine if it should encrypt the data, so Hadoop security
WILL break down in this case.